Add files via upload
This paper, "Does Big Data really need to be Big?" , is a final report for the COMP 8157 Advanced Database Topics course at the University of Windsor. The work was submitted on March 26, 2025, by Group 01, "Big Hero 06". The student contributors are Arshnoor Singh Sohi, Ashiqur Rahman, Gagandeep Kaur, Foysal Rahman Nitu, Tausif Zaman, and Yingying Lin.

The research explores the concept of "dark data," which is data collected by organizations but not used for analytics. It notes that 50% to 90% of collected data can be classified as dark. The paper questions if the vastness of big data is necessary and if the same analytical results can be achieved with reduced datasets.

To address these questions, the authors propose the Quality-Driven Data Reduction (QDDR) framework. They also developed a Dark Data Transformation Module (DDTM) to convert various data formats like JSON, H5, and TXT into analyzable CSV structures. Using synthetic healthcare data, the project demonstrates a pipeline to transform and give visibility to this unused data.

The paper acknowledges challenges, including the difficulty of finding real-world dark datasets and the potential costs for businesses to invest in unearthing data that isn't guaranteed to be valuable. Future work suggests integrating machine learning, validating with real-world data, developing mechanisms for handling sensitive information, and exploring the use of Large Language Models (LLMs).